adp/requirements.txt
requests==2.9.1
schedule==0.4.2
pymongo==3.4.0
beautifulsoup4==4.5.3
paho_mqtt==1.2.3
